layering: the neutral layers ask the rail instead of naming it
Three leaks where a neutral module reached sideways into one rail's code.

**`model` called `draw::vulkan::forget_plane_draw_ring`.** The plane draw ring
is keyed by mapping id and must be dropped when that id stops naming its
surface, or a recycled id inherits its predecessor's passes — a wrong reading of
a live plane. The obligation is the model's; the record is the rail's. It
becomes `Backend::forget_plane_draws`, the companion to the existing
`plane_draw_witness`, whose empty-string default already states that a rail may
keep no such record.

**`runtime::icb::clear_icb_cache` called `icb::metal::clear_host_icb_cache`.**
Its own doc states the invariant — one entry point clears both maps, because a
registry entry outliving its host object names a descriptor nothing was built
from. That stays true and stops being spelled as a `cfg`:
`Backend::forget_host_icbs` is the rail's half, and `MetalBackend::reset` no
longer reaches into the runtime for it.

**`compute_exec::multisample_sampled_texture` lived in the neutral module.** It
is `#[cfg(feature = "backend-vulkan")]`, returns `StagedTexture<VulkanStage>`,
has one caller in `compute_exec::vulkan`, and resolves its target through
`draw::vulkan::gva_span_identity`. It was Vulkan code in a neutral file; it
moves to `compute_exec/vulkan.rs`, where the `cfg`, the `vulkan::` qualifiers
and the cross-module path all collapse.

`compute_exec/mod.rs` now names no rail outside the `ComputeStatus::MetalBackend`
variant.

Verified: feature matrix all arms, 2767 passed / 0 failed, clippy clean on all
three arms (only the four pre-existing dead-code warnings remain).
